Development enemy animations and states

+++ Assets/Scripts/Core/Interfaces/IDamageable.cs
@@ -5,6 +5,6 @@ namespace Assets.Scripts.Core.Interfaces
 {
     public interface IDamageable
     {
-        public void TakeDamage(float damageValue);
+        public void TakeDamage(int damageValue);
     }
 }
